Fixed bug when importing users by script: new users have access to all other calendars MERGE FROM PATCHS-6-7-0

+++ utilit/usermodifyincl.php
@@ -156,23 +156,25 @@ class bab_userModify {
 				'".$id."'
 				)");
 	
-			$iDefaultCalendarSiteAccess = (int)($babBody->babsite['iDefaultCalendarAccess']);
-			if( $iDefaultCalendarSiteAccess == BAB_CAL_ACCESS_NONE )
+			if( isset($babBody->babsite) && isset($babBody->babsite['iDefaultCalendarAccess'] ))
 				{
-				$babDB->db_query("insert into ".BAB_CALACCESS_USERS_TBL." (id_cal, id_user, bwrite) select c.id, ".$id.", uo.iDefaultCalendarAccess from ".BAB_CALENDAR_TBL." c left join ".BAB_CAL_USER_OPTIONS_TBL." uo on c.owner=uo.id_user where type=".BAB_CAL_USER_TYPE." and uo.iDefaultCalendarAccess is not null and uo.iDefaultCalendarAccess != ".BAB_CAL_ACCESS_NONE);
-				}
-			else
-				{
-				/* hack MySQL : Because we can't select rows without iDefaultCalendarAccess = -1, so we add them with 100 as bwrite value and then delete them */
-				$babDB->db_query("insert into ".BAB_CALACCESS_USERS_TBL." (id_cal, id_user, bwrite) select c.id, ".$id.", if(uo.iDefaultCalendarAccess = -1, 100, if(uo.iDefaultCalendarAccess is null, ".$iDefaultCalendarSiteAccess.", uo.iDefaultCalendarAccess))  from ".BAB_CALENDAR_TBL." c left join ".BAB_CAL_USER_OPTIONS_TBL." uo on c.owner=uo.id_user where type=".BAB_CAL_USER_TYPE."");
-				$babDB->db_query("delete from ".BAB_CALACCESS_USERS_TBL." where bwrite = 100");
-				}
+				$iDefaultCalendarSiteAccess = (int)($babBody->babsite['iDefaultCalendarAccess']);
+				if( $iDefaultCalendarSiteAccess == BAB_CAL_ACCESS_NONE )
+					{
+					$babDB->db_query("insert into ".BAB_CALACCESS_USERS_TBL." (id_cal, id_user, bwrite) select c.id, ".$id.", uo.iDefaultCalendarAccess from ".BAB_CALENDAR_TBL." c left join ".BAB_CAL_USER_OPTIONS_TBL." uo on c.owner=uo.id_user where type=".BAB_CAL_USER_TYPE." and uo.iDefaultCalendarAccess is not null and uo.iDefaultCalendarAccess != ".BAB_CAL_ACCESS_NONE);
+					}
+				else
+					{
+					/* hack MySQL : Because we can't select rows without iDefaultCalendarAccess = -1, so we add them with 100 as bwrite value and then delete them */
+					$babDB->db_query("insert into ".BAB_CALACCESS_USERS_TBL." (id_cal, id_user, bwrite) select c.id, ".$id.", if(uo.iDefaultCalendarAccess = -1, 100, if(uo.iDefaultCalendarAccess is null, ".$iDefaultCalendarSiteAccess.", uo.iDefaultCalendarAccess))  from ".BAB_CALENDAR_TBL." c left join ".BAB_CAL_USER_OPTIONS_TBL." uo on c.owner=uo.id_user where type=".BAB_CAL_USER_TYPE."");
+					$babDB->db_query("delete from ".BAB_CALACCESS_USERS_TBL." where bwrite = 100");
+					}
 
-			if( $iDefaultCalendarSiteAccess != BAB_CAL_ACCESS_NONE )
-				{
-				$babDB->db_query("insert into ".BAB_CALACCESS_USERS_TBL." (id_cal, id_user, bwrite) select ".$idusercal.", id, ".$iDefaultCalendarSiteAccess." from ".BAB_USERS_TBL." where id !='".$id."'");
+				if( $iDefaultCalendarSiteAccess != BAB_CAL_ACCESS_NONE )
+					{
+					$babDB->db_query("insert into ".BAB_CALACCESS_USERS_TBL." (id_cal, id_user, bwrite) select ".$idusercal.", id, ".$iDefaultCalendarSiteAccess." from ".BAB_USERS_TBL." where id !='".$id."'");
+					}
 				}
-
 
 			if( $bgroup && isset($babBody->babsite['idgroup']) && $babBody->babsite['idgroup'] != 0)
 				{
